Extract `FeedImageCell` test helpers to a separate file
