+++ EssentialFeediOSTests/Helpers/FeedImageCell+TestHelpers.swift
@@ -1,0 +1,40 @@
+//
+//  FeedImageCell+TestHelpers.swift
+//  EssentialFeediOSTests
+//
+//  Created by Christos Tzimas on 28/5/25.
+//
+
+import EssentialFeediOS
+import Foundation
+
+extension FeedImageCell {
+    
+    func simulateRetryAction() {
+        feedImageRetryButton.simulateTap()
+    }
+    
+    var isShowingLocation: Bool {
+        !locationContainer.isHidden
+    }
+    
+    var isShowingImageLoadingIndicator: Bool {
+        feedImageContainer.isShimmering
+    }
+    
+    var isShowingRetryAction: Bool {
+        !feedImageRetryButton.isHidden
+    }
+    
+    var locationText: String? {
+        locationLabel.text
+    }
+    
+    var descriptionText: String? {
+        descriptionLabel.text
+    }
+    
+    var renderedImage: Data? {
+        feedImageView.image?.pngData()
+    }
+}
